Don't add include/ to the includes by default (proposed fix for #645)
When compiling for Linux or macOS, we have system wide installed headers and
libraries that are usually newer than the ones in the include/ directory.
Their inclusion is likely to break builds. For that reason an if clause has
been added that just adds the include/ directory to the compiler includes when
building for Windows.